no errors, 12 failures

+++ src/test/java/com/qa/ims/persistence/dao/ItemDAOTest.java
@@ -3,13 +3,13 @@ package com.qa.ims.persistence.dao;
 import static org.junit.Assert.assertEquals;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import org.junit.Before;
 import org.junit.Test;
 
 import com.qa.ims.persistence.domain.Item;
 import com.qa.ims.utils.DBUtils;
-import com.sun.tools.javac.util.List;
 
 public class ItemDAOTest {
 	
@@ -23,31 +23,31 @@ public class ItemDAOTest {
 	
 	@Test
 	public void testCreate() {
-		final Item created = new Item(2L, "bird_table", 40.00);
+		final Item created = new Item(2L, "bird_table", (double) 40);
 		assertEquals(created, DAO.create(created));
 	}
 	
 	@Test
 	public void testReadAll() {
 		List<Item> expected = new ArrayList<>();
-		expected.add( new Item(1L, "plant_pot", 20.00));
+		expected.add( new Item(1L, "plant_pot", (double) 20));
 		assertEquals(expected, DAO.readAll());
 	}
 	
 	@Test
 	public void testReadLatest() {
-		assertEquals(new Item(1L, "plant_pot", 20.00), DAO.readLatest());
+		assertEquals(new Item(1L, "plant_pot", (double) 20), DAO.readLatest());
 	}
 	
 	@Test
 	public void testRead() {
 		final long ID = 1L;
-		assertEquals(new Item(ID, "plant_pot", 20.00), DAO.read(ID));
+		assertEquals(new Item(ID, "plant_pot", (double) 20), DAO.read(ID));
 	}
 	
 	@Test
 	public void testUpdate() {
-		final Item updated = new Item(1L, "bird_table", 40.00);
+		final Item updated = new Item(1L, "bird_table", (double) 40);
 		assertEquals(updated, DAO.update(updated));
 	}
